final tweaks to get feat-adjustments in so we can support feats on objects
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@4728 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ structs.h
@@ -2702,7 +2702,7 @@ struct obj_flag_data {
  * and already have a player base and don't want to do a player wipe. */
 struct obj_affected_type {
     byte location; /**< Which ability to change (APPLY_XXX) */
-    sbyte modifier; /**< How much it changes by              */
+    int modifier; /**< How much it changes by              */
     int bonus_type; /**< What type of bonus is this. */
 };
